raw: DataSet: simplify __str__()
Don't print out the raw data as it can be quite huge. I think this data
is too much to be outputted on the command line by
default for all users.

One example on my system is a .raw file which is 2MB. I don't see much
value in printing it on the command line.